Corrected return type.

+++ src/Magick.NET/Native/MagickImage.cs
@@ -24,7 +24,7 @@ namespace ImageMagick
         [UnmanagedFunctionPointer(CallingConvention.Cdecl)]
         private delegate bool ProgressDelegate(IntPtr origin, long offset, ulong extent, IntPtr userData);
         [UnmanagedFunctionPointer(CallingConvention.Cdecl)]
-        private delegate int ReadWriteStreamDelegate(IntPtr data, UIntPtr length, IntPtr user_data);
+        private delegate long ReadWriteStreamDelegate(IntPtr data, UIntPtr length, IntPtr user_data);
         [UnmanagedFunctionPointer(CallingConvention.Cdecl)]
         private delegate long SeekStreamDelegate(long offset, IntPtr whence, IntPtr user_data);
         [UnmanagedFunctionPointer(CallingConvention.Cdecl)]
